Mechanism to configure threaded library use.
Added options to the setup.cfg file allowing to specify if threaded
versions of the fftw and nfft libraries are to be linked in. This
allows the installations of pyNFFT on platforms where no threaded
fftw or nfft libraries are available (e.g., the native compiler
Mac OS-X (including 10.9.x) does not support openmp, so nfft can't
be build in a threaded version).