First working version with pybind11
The C structs for NFFT plans and FFTW complex types
are now wrapped in a templated type, which makes them
easy to use directly, e.g., as class members in the
_NFFT class template.

Further, all C functions using name mangling are now
wrapped minimally with function templates, with the
same effect as above.

Finally, precompute() is now implemented.